Move pill parsing into deserialize and avoid intermediate objects.
